requirements.txt
flask>=2.2.0,<3.0
crewai==0.121.1
pydantic>=2.4.2        # Avoid Pydantic V2 breaking changes for now
python-dotenv>=1.0.0           # For environment variable loading
requests>=2.28.0               # Common for API calls if needed
gunicorn>=20.1.0               # For production deployment WSGI server
click>=8.1.3                  # Flask dependency but pinned for safety
typing_extensions>=4.5.0       # Compatibility for Python typing
lxml>=4.9.3
Jinja2>=3.1.2
crewai-tools[mcp]
langchain-google-genai
phonenumbers
google-api-python-client
chromadb
tavily-python
duckduckgo_search
